[CHORE] #60 - 댓글삭제 api CommentRepository로 이동
