+++ app/src/main/java/com/fork/spoonfeed/data/remote/datasource/CommentDataSourceImpl.kt
@@ -2,11 +2,17 @@ package com.fork.spoonfeed.data.remote.datasource
 
 import com.fork.spoonfeed.data.remote.api.community.CommentService
 import com.fork.spoonfeed.data.remote.model.community.RequestCommentData
+import com.fork.spoonfeed.data.remote.model.community.RequestDeleteCommentData
 import com.fork.spoonfeed.data.remote.model.community.ResponseCommentData
+import com.fork.spoonfeed.data.remote.model.community.ResponseDeleteCommentData
 
 class CommentDataSourceImpl(private val commentService: CommentService) : CommentDataSource {
 
     override suspend fun postComment(pk: Int, body: RequestCommentData): ResponseCommentData {
         return commentService.postComment(pk = pk, body = body)
+    }
+
+    override suspend fun deleteComment(pk: Int, body: RequestDeleteCommentData): ResponseDeleteCommentData {
+        return commentService.deleteComment(pk = pk, body = body)
     }
 }
