+++ app/src/main/java/com/fork/spoonfeed/data/remote/datasource/PostDataSource.kt
@@ -16,6 +16,4 @@ interface PostDataSource {
     suspend fun patchPost(pk: Int, body: RequestPatchPostData): ResponsePatchPostData
 
     suspend fun deletePost(pk: Int): ResponseDeletePostData
-
-    suspend fun deleteComment(pk: Int, body: RequestDeleteCommentData): ResponseDeleteCommentData
 }
